+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * added active class to menu links
 * added dropdown to menu
+* added two-way binding example
 
 ## 0.0.7 (2016-01-13)